Prototype architecture finished.
