+++ DatabaseBenchmark.Core/Benchmarking/ITest.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DatabaseBenchmark.Core.Benchmarking
+{
+    public interface ITest
+    {
+        void StartTest();
+        void StopTest();
+    }
+}
