+++ DatabaseBenchmark.Core/Benchmarking/TestFactory.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DatabaseBenchmark.Core.Benchmarking
+{
+    public static class TestFactory
+    {
+        public static ITest CreateTest(Type testType, params object[] parameters)
+        {
+            return null;
+        }
+    }
+}
