--- DatabaseBenchmark.Core/Report/ReportEnums.cs
@@ -1,16 +1,0 @@
-﻿
-namespace DatabaseBenchmark.Report
-{
-    public enum ReportType : byte
-    {
-        Summary,
-        Detailed
-    }
-
-    public enum ReportFormat : byte
-    {
-        CSV,
-        JSON,
-        PDF
-    }
-}
