+++ DatabaseBenchmark.Core/Reporting/ReportEnums.cs
@@ -1,0 +1,16 @@
+﻿
+namespace DatabaseBenchmark.Reporting
+{
+    public enum ReportType : byte
+    {
+        Summary,
+        Detailed
+    }
+
+    public enum ReportFormat : byte
+    {
+        CSV,
+        JSON,
+        PDF
+    }
+}
